Splitting basic_action and transfer_action into their own headers
